requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Azure OpenAI
openai==1.3.7

# LangChain and LangSmith
langchain>=0.3.0
langchain-openai>=0.1.0
langgraph==0.0.20
langsmith>=0.1.17

# Image processing
Pillow==10.1.0

# PDF processing
tika==2.6.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0